Final fixes fo the night

+++ idea-service/src/projects/projects.service.ts
@@ -1,10 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Project } from '../entities/project.entity';
 import { Administers } from '../entities/administers.entity';
 import { Member } from 'src/entities/member.entity';
 import { Idea } from 'src/entities/idea.entity';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class ProjectService {
@@ -13,7 +14,7 @@ export class ProjectService {
     @InjectRepository(Administers) private adminRepository: Repository<Administers>,
     @InjectRepository(Member) private memberRepository: Repository<Member>,
     @InjectRepository(Idea) private ideaRepository: Repository<Idea>,
-    
+    @InjectRepository(User) private userRepository: Repository<User>,
     ) {}
   
     /**
@@ -21,28 +22,38 @@ export class ProjectService {
      * @param userId The id of the user
      */
   async getAdminProjects(userId: number): Promise<any> {
-    const projectIds = await this.adminRepository.find({userId});
-    const projects = [];
-    for (let i = 0; i < projectIds.length; i++) {
-      const id = projectIds[i].projectId;
-      const project = await this.projectRepository.findOne({ projectId: id });
-      projects.push(project);
+    const user = await this.userRepository.findOne({userId});
+    if (user) {
+      const projectIds = await this.adminRepository.find({userId});
+      const projects = [];
+      for (let i = 0; i < projectIds.length; i++) {
+        const id = projectIds[i].projectId;
+        const project = await this.projectRepository.findOne({ projectId: id });
+        projects.push(project);
+      }
+      return projects;
+    } else {
+      throw new HttpException('UserId not found', HttpStatus.NOT_FOUND);
     }
-    return projects;
   }
   /**
    * Get the projects that a given user is a member of
    * @param userId The id of the user
    */
   async getUserProjects(userId: number): Promise<any> {
-    const projectIds = await this.memberRepository.find({ userId });
-    const projects = [];
-    for (let i = 0; i < projectIds.length; i++) {
-      const id = projectIds[i].projectId;
-      const project = await this.projectRepository.findOne({ projectId: id });
-      projects.push(project);
+    const user = await this.userRepository.findOne({userId});
+    if (user) {
+      const projectIds = await this.memberRepository.find({ userId });
+      const projects = [];
+      for (let i = 0; i < projectIds.length; i++) {
+        const id = projectIds[i].projectId;
+        const project = await this.projectRepository.findOne({ projectId: id });
+        projects.push(project);
+      }
+      return projects;
+    } else {
+      throw new HttpException('UserId not found', HttpStatus.NOT_FOUND);
     }
-    return projects;
   }
   /**
    * Retrieves a specific project
@@ -50,7 +61,11 @@ export class ProjectService {
    */
   async getProject(projectId: string): Promise<any> {
     const project = await this.projectRepository.findOne(projectId);
-    return project;
+    if (project) {
+      return project;
+    } else {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND);
+    }
   }
   /**
    * Creates a project
@@ -60,26 +75,40 @@ export class ProjectService {
    * @param color The color that the user has assigned to the project
    */
   async createProject(userId: number, name: string, description: string, color: string): Promise<any> {
-    const project = await this.projectRepository.insert({ name, description, color});
-    const projectId = project.identifiers[0].projectId
-    this.adminRepository.insert({userId, projectId});
-    return;
+    if (name && description && color) {
+      const project = await this.projectRepository.insert({ name, description, color});
+      const projectId = project.identifiers[0].projectId
+      this.adminRepository.insert({userId, projectId});
+      return projectId;
+    } else {
+      throw new HttpException('Name, description, and color must be defined', HttpStatus.BAD_REQUEST);
+    }
   }
   /**
    * Adds a user as a member of a project
    * @param userId The id of the user to add to the project
    * @param projectId The id of the project to add the user to
    */
-  addUserToProject(userId: number, projectId: number): any {
-    this.memberRepository.insert({userId, projectId});
+  async addUserToProject(userId: number, projectId: number): Promise<any> {
+    const user = await this.userRepository.findOne({userId});
+    if (user) {
+      return this.memberRepository.insert({userId, projectId});
+    } else {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
   }
   /**
    * Adds a user as an admin of a project
    * @param userId The id of the user to add as an admin to a project
    * @param projectId The id of the project to add the user to
    */
-  addAdminToProject(userId: number, projectId: number): any {
-    this.adminRepository.insert({userId, projectId});
+  async addAdminToProject(userId: number, projectId: number): Promise<any> {
+    const user = await this.userRepository.findOne({userId});
+    if (user) {
+      this.adminRepository.insert({userId, projectId});
+    } else {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
   }
   /**
    * 
@@ -88,31 +117,51 @@ export class ProjectService {
    * @param description The description of the project to be updated
    * @param color The color of the project to be updated
    */
-  updateProject(projectId: number, name: string, description: string, color: string): any {
-    this.projectRepository.update(projectId, {name, description, color});
+  async updateProject(projectId: number, name: string, description: string, color: string): Promise<any> {
+    const project = await this.projectRepository.findOne(projectId);
+    if (project) {
+      return this.projectRepository.update(projectId, {name, description, color});
+    } else {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND);
+    }
   }
   /**
    * Removes a user from the project
    * @param userId The id of the user to delete from the project
    * @param projectId The id of the project to delete the user from
    */
-  deleteUsersFromProject(userId: number, projectId: number): any {
-    this.memberRepository.delete({ userId, projectId })
+  async deleteUsersFromProject(userId: number, projectId: number): Promise<any> {
+    const user = await this.userRepository.findOne({userId});
+    if (user) {
+      this.memberRepository.delete({ userId, projectId })
+    } else {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
   }
   /**
    * Gets all the given ideas for a project
    * @param projectId The projectId to get the ideas for
    */
-  getAllIdeasForProject(projectId: number): any {
-    this.ideaRepository.createQueryBuilder('Idea').where('projectId = :projectId', { projectId: projectId }).getMany();
+  async getAllIdeasForProject(projectId: number): Promise<any> {
+    const project = await this.projectRepository.findOne(projectId);
+    if (project) {
+      this.ideaRepository.createQueryBuilder('Idea').where('projectId = :projectId', { projectId: projectId }).getMany();
+    } else {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND);
+    }
   }
   /**
    * Delete a project as well as it's admin and member references
    * @param projectId The projectId of the project to get deleted
    */
-  deleteProject(projectId: number): any {
-    this.projectRepository.delete(projectId);
-    this.memberRepository.delete(projectId);
-    this.adminRepository.delete(projectId)
+  async deleteProject(projectId: number): Promise<any> {
+    const project = await this.projectRepository.findOne(projectId);
+    if (project) {
+      this.projectRepository.delete(projectId);
+      this.memberRepository.delete({projectId: projectId});
+      this.adminRepository.delete({projectId: projectId});
+    } else {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND);
+    }
   }
 }
